bert/requirements.txt
torch==1.7.1+cu101
torchaudio==0.7.2
torchvision==0.8.2+cu101
pytorch-crf
sklearn
transformers